+++ girder_annotation/girder_large_image_annotation/rest/annotation.py
@@ -46,18 +46,20 @@ class AnnotationResource(Resource):
 
         self.resourceName = 'annotation'
         self.route('GET', (), self.find)
+        self.route('POST', (), self.createAnnotation)
         self.route('GET', ('schema',), self.getAnnotationSchema)
         self.route('GET', ('images',), self.findAnnotatedImages)
         self.route('GET', (':id',), self.getAnnotation)
-        self.route('GET', (':id', 'history'), self.getAnnotationHistoryList)
-        self.route('GET', (':id', 'history', ':version'), self.getAnnotationHistory)
-        self.route('PUT', (':id', 'history', 'revert'), self.revertAnnotationHistory)
-        self.route('POST', (), self.createAnnotation)
-        self.route('POST', (':id', 'copy'), self.copyAnnotation)
         self.route('PUT', (':id',), self.updateAnnotation)
         self.route('DELETE', (':id',), self.deleteAnnotation)
         self.route('GET', (':id', 'access'), self.getAnnotationAccess)
         self.route('PUT', (':id', 'access'), self.updateAnnotationAccess)
+        self.route('POST', (':id', 'copy'), self.copyAnnotation)
+        self.route('GET', (':id', 'history'), self.getAnnotationHistoryList)
+        self.route('GET', (':id', 'history', ':version'), self.getAnnotationHistory)
+        self.route('PUT', (':id', 'history', 'revert'), self.revertAnnotationHistory)
+        self.route('PUT', (':id', 'metadata'), self.setMetadata)
+        self.route('DELETE', (':id', 'metadata'), self.deleteMetadata)
         self.route('GET', ('item', ':id'), self.getItemAnnotations)
         self.route('POST', ('item', ':id'), self.createItemAnnotations)
         self.route('DELETE', ('item', ':id'), self.deleteItemAnnotations)
@@ -66,9 +68,9 @@ class AnnotationResource(Resource):
         self.route('GET', ('folder', ':id', 'create'), self.canCreateFolderAnnotations)
         self.route('PUT', ('folder', ':id', 'access'), self.setFolderAnnotationAccess)
         self.route('DELETE', ('folder', ':id'), self.deleteFolderAnnotations)
+        self.route('GET', ('counts',), self.getItemListAnnotationCounts)
         self.route('GET', ('old',), self.getOldAnnotations)
         self.route('DELETE', ('old',), self.deleteOldAnnotations)
-        self.route('GET', ('counts',), self.getItemListAnnotationCounts)
 
     @describeRoute(
         Description('Search for annotations.')
@@ -105,7 +107,8 @@ class AnnotationResource(Resource):
             query['annotation.name'] = params['name']
         fields = list(
             (
-                'annotation.name', 'annotation.description', 'access', 'groups', '_version'
+                'annotation.name', 'annotation.description', 'annotation.attributes',
+                'access', 'groups', '_version'
             ) + Annotation().baseFields)
         return Annotation().findWithPermissions(
             query, sort=sort, fields=fields, user=self.getCurrentUser(),
@@ -793,3 +796,47 @@ class AnnotationResource(Resource):
                     results['referenced'] = {}
                 results['referenced'][itemId] = True
         return results
+
+    @access.user(scope=TokenScope.DATA_WRITE)
+    @filtermodel(model='annotation', plugin='large_image')
+    @autoDescribeRoute(
+        Description('Set metadata (annotation.attributes) fields on an annotation.')
+        .responseClass('Annotation')
+        .notes('Set metadata fields to null in order to delete them.')
+        .param('id', 'The ID of the annotation.', paramType='path')
+        .jsonParam('metadata', 'A JSON object containing the metadata keys to add',
+                   paramType='body', requireObject=True)
+        .param('allowNull', 'Whether "null" is allowed as a metadata value.', required=False,
+               dataType='boolean', default=False)
+        .errorResponse(('ID was invalid.',
+                        'Invalid JSON passed in request body.',
+                        'Metadata key name was invalid.'))
+        .errorResponse('Write access was denied for the annotation.', 403)
+    )
+    @loadmodel(model='annotation', plugin='large_image', getElements=False, level=AccessType.WRITE)
+    def setMetadata(self, annotation, metadata, allowNull):
+        return Annotation().setMetadata(annotation, metadata, allowNull=allowNull)
+
+    @access.user(scope=TokenScope.DATA_WRITE)
+    @filtermodel(model='annotation', plugin='large_image')
+    @autoDescribeRoute(
+        Description('Delete metadata (annotation.attributes) fields on an annotation.')
+        .responseClass('Item')
+        .param('id', 'The ID of the annotation.', paramType='path')
+        .jsonParam(
+            'fields', 'A JSON list containing the metadata fields to delete',
+            paramType='body', schema={
+                'type': 'array',
+                'items': {
+                    'type': 'string'
+                }
+            }
+        )
+        .errorResponse(('ID was invalid.',
+                        'Invalid JSON passed in request body.',
+                        'Metadata key name was invalid.'))
+        .errorResponse('Write access was denied for the annotation.', 403)
+    )
+    @loadmodel(model='annotation', plugin='large_image', getElements=False, level=AccessType.WRITE)
+    def deleteMetadata(self, annotation, fields):
+        return Annotation().deleteMetadata(annotation, fields)